agentic-knowledge-system/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langgraph==0.2.0
langchain-core==0.3.0
langchain-groq==0.2.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.26.0
